+++ frontend/src/components/Footer.js
@@ -1,0 +1,5 @@
+function Footer() {
+ return <div>Not Found</div>;
+}
+
+export default Footer;
